% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcut.R
\name{boxcut}
\alias{boxcut}
\title{Breaks a vector as a boxplot}
\usage{
boxcut(x, hinge = 1.5)
}
\arguments{
\item{x}{vector}

\item{hinge}{scalar}
}
\value{
vector
}
\description{
Breaks a vector as a boxplot
}
