% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{Outersection of two vectors}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector
}
\description{
https://www.r-bloggers.com/outersect-the-opposite-of-rs-intersect-function/
}
\examples{
outersect(c(1, 2, 3), c(2, 3, 4))
}
