% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_words.R
\name{str_extract_words}
\alias{str_extract_words}
\title{str_extract_words}
\usage{
str_extract_words(string, n_words)
}
\arguments{
\item{n_words}{}
}
\value{
string
}
\description{
str_extract_words
}
